requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.1
snowflake-connector-python==3.12.2
python-dotenv==1.0.1
numpy==1.26.4